Add calendario_de_riego
This device will not start the program
if the day has been set as not "R" rain worthy.

it works by checking the day of the week.
it compares to what the file
calendario_de_riego.json has and if it is
not a R, then it does not start.