addded a way to pick a different index depending on state
when running a program, we only need to be able to
cancel it. not start a new manual run

when not running a program. only able to
pick a program to run.

also updated:
cancelled: state to cancel a runnning program
riego suspendido: act of blocking the chance to
run a program for X days.

TODO: updating of the index page does not work
after doing an action.